requirements.txt
streamlit>=1.0
sqlalchemy>=1.4
psycopg2-binary>=2.8
pandas>=1.4
altair>=4.0
pytz>=2021.3
python-dotenv>=0.19
streamlit-autorefresh>=0.1
